requirements.txt
requests
python-dotenv
scikit-learn
pandas
joblib
beautifulsoup4
lxml
